requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
onnxruntime>=1.16.0
whisper-openai
transformers>=4.30.0
hyperpyyaml>=1.0.0
pyyaml
